extract new detector_utils file from entropy keyword combinator
